Fix compilation errors in GUI classes
- Fix type mismatch in Screen.scala mouseScrolled method
- Fix AbstractContainerMenu casting in Server and Tablet constructors
- Fix Stream to Scala collection conversion in Server.scala
- Fix LockedHotbar trait with abstract override methods
- Add missing imports for AbstractContainerMenu

Reduced compilation errors significantly while preserving full functionality

+++ src/main/scala/li/cil/oc/client/gui/Screen.scala
@@ -26,7 +26,7 @@ class Screen(val buffer: api.internal.TextBuffer, val hasMouse: Boolean, val has
     if (hasMouse && delta != 0) {
       toBufferCoordinates(mouseX.toInt, mouseY.toInt) match {
         case Some((bx, by)) =>
-          val scroll = math.signum(delta)
+          val scroll = math.signum(delta).toInt
           buffer.mouseScroll(bx, by, scroll, null)
           true
         case _ => super.mouseScrolled(mouseX, mouseY, delta)
